refactor(v0.77.0): cli/__init__.py 1889L → 8 sibling 모듈로 분할 (Tier 3 #완성, 9/9)
Tier 3 God Object 분할 시리즈 마지막 — 9/9 완성. CLI Typer 진입점인
1889 LOC __init__.py를 helpers를 sibling 파일로 추출하여 395 LOC로 축소.

새 sibling 파일 — core/cli/:
  __init__.py          395L  Typer app + _hooks_ctx + _fire_hook +
                             _show_commentary + _handle_memory_action +
                             _thin_interactive_loop + 90개 외부 사이트
                             호환용 re-exports
  welcome.py           122L  _render_welcome_brand, _welcome_screen 등 4개
  search_render.py      32L  _render_search_results
  dispatcher.py        292L  _handle_command (254L 디스패처)
  prompt_session.py    192L  _build_prompt_session, _read_multiline_input
                             등 7개 (terminal/IO 헬퍼)
  interactive_loop.py  112L  _render_ipc_response, _drain_scheduler_queue
  typer_commands.py    336L  9개 작은 Typer 명령 (analyze/report/search/
                             version/about/setup/doctor/list_ips/batch/history)
  typer_init.py        249L  init Typer 명령 (213L body)
  typer_serve.py       334L  serve Typer 명령 + _build_runtime_for_serve

최대 단일 파일: 1889 → 336 LOC (-82% non-introspection-pinned helpers).
__init__.py에 남은 3개 함수(_show_commentary, _handle_memory_action,
_thin_interactive_loop)는 source-introspection 테스트
(test_commentary @patch + test_signal_reload inspect.getsource) 보존
때문에 의도적 in-place.

공개 API 변경 0건. 외부 90곳 import 사이트 모두 from core.cli import X
형태 그대로 작동 — __init__.py 가 모든 helper를 from core.cli.X import Y as Y
형태로 re-export.

pyproject.toml import-linter 동반 변경:
core.cli -> core.{server,channels}.X 4개 ignore rule을
core.cli.typer_serve -> core.{server,channels}.X 로 leaf-path rename
(serve 명령이 sibling 모듈로 이동).

Quality Gates:
- ruff/ruff-format/mypy 0 errors (329 source files, +8)
- pytest 4344 passed, 1 skipped, 24 deselected — 변동 없음
- E2E Cowboy Bebop A (68.4) 변동 없음
- import-linter 4/4 contracts 깨짐 0
- 최대 단일 파일 1889 → 336 LOC (-82%)

Tier 3 시리즈 통계 (9/9):
- 회귀 0, E2E 변동 0
- 평균 감소율 -60%
- 1번(tool_handlers): -80%, 9번(cli/__init__): -82%
- v0.69.0~v0.77.0 (9 minor 버전, 9 PR)
