+++ core/cli/search_render.py
@@ -1,0 +1,32 @@
+"""Search result rendering for the GEODE thin CLI.
+
+Extracted from ``core/cli/__init__.py`` (Tier 3 God Object split).
+"""
+
+from __future__ import annotations
+
+from typing import Any
+
+from core.ui.console import console
+
+
+def _render_search_results(query: str, results: list[Any]) -> None:
+    """Render IP search results."""
+    console.print()
+    console.print(f"  [header]Search: '{query}'[/header]")
+
+    if not results:
+        console.print("  [muted]No matching IPs found.[/muted]")
+        console.print()
+        return
+
+    for r in results:
+        bar_len = int(r.score * 20)
+        bar = "█" * bar_len + "░" * (20 - bar_len)
+        style = "success" if r.score >= 0.5 else "muted"
+        console.print(
+            f"    [{style}]{bar}[/{style}] [dim]relevance[/dim] {r.score:.0%}"
+            f"  [value]{r.ip_name}[/value]"
+        )
+        console.print(f"      [muted]matched: {', '.join(r.matches[:5])}[/muted]")
+    console.print()
